requirements.txt
Django~=5.0.6
psycopg2~=2.9.9
python-dotenv~=1.0.1
ruff~=0.4.5
pip~=24.0
requests~=2.32.3